CART - function part 3

+++ resources/functions.php
@@ -1,36 +1,56 @@
 <?php 
 
 function cart() {
-  $query = query("SELECT * FROM products");
-  confirm($query);
 
-  while($row = fetch_array($query)) {
-    ?>
-    <tr>
-      <td><?= $row['title'] ?></td>
-      <td>$ <?= $row['price'] ?></td>
-      <td><?= $row['quantity'] ?></td>
-      <td><?= $row['quantity'] ?></td>
-      <td>
-      	<!-- remopve apenas um  -->
-      	<a class="btn btn-sm btn-warning" href="cart.php?remove=<?= $row['id'] ?>">
-      		<span class="glyphicon glyphicon-minus" title="Remove one item <?= $row['title'] ?>"></span>		
-      	</a>
+	// pegando o id correto
+	foreach ($_SESSION as $name => $value) {
 
-      	<!-- adiciona -->
-      	<a class="btn btn-sm btn-success" href="cart.php?add=<?= $row['id'] ?>">
-      		<span class="glyphicon glyphicon-plus" title="Add one more item <?= $row['title'] ?>"></span>
-      	</a>
+		
+		echo substr($name, 0, 8);
+		// die();
+		
+		
+		if (substr($name, 0, 8) == 'product_') {
+			
+		  $query = query("SELECT * FROM products");
+		  confirm($query);
 
-      	<!-- exclui todos -->
-      	<a title="Remove all items <?= $row['title'] ?>" class="btn btn-sm btn-danger" href="cart.php?delete=<?= $row['id'] ?>">
-      		<span class="glyphicon glyphicon-trash"></span>      			
-      	</a>
-      </td>                  
-    </tr> 
-    <?php
-  }
-}  
+		  while($row = fetch_array($query)) {
+		    ?>
+		    <tr>
+		      <td><?= $row['title'] ?></td>
+		      <td>$ <?= $row['price'] ?></td>
+		      <td><?= $row['quantity'] ?></td>
+		      <td><?= $row['quantity'] ?></td>
+		      <td>
+		      	<!-- remopve apenas um  -->
+		      	<a class="btn btn-sm btn-warning" href="cart.php?remove=<?= $row['id'] ?>">
+		      		<span class="glyphicon glyphicon-minus" title="Remove one item <?= $row['title'] ?>"></span>		
+		      	</a>
+
+		      	<!-- adiciona -->
+		      	<a class="btn btn-sm btn-success" href="cart.php?add=<?= $row['id'] ?>">
+		      		<span class="glyphicon glyphicon-plus" title="Add one more item <?= $row['title'] ?>"></span>
+		      	</a>
+
+		      	<!-- exclui todos -->
+		      	<a title="Remove all items <?= $row['title'] ?>" class="btn btn-sm btn-danger" href="cart.php?delete=<?= $row['id'] ?>">
+		      		<span class="glyphicon glyphicon-trash"></span>      			
+		      	</a>
+		      </td>                  
+		    </tr> 
+		    <?php
+
+		  } // end while
+
+		} // end if subtr()
+
+	} // end foreach()
+
+} // end function cart()  
+
+
+
 
 function set_message($message) {
 	if (!empty($message)) {
